Add toast module for notifications
Shows user a toast after a notification, i.e., an account being
successfully added or an error. Hides itself after 8 seconds.
Transitions are also nice. One issue, however, is there's no way to
dismiss it instead of waiting the 8 seconds. Particularly because, it a
user hovers over the countdown stops... anyways. It's short and sweet.