Persist window state after app is closed
Listen to resize, move, maximize and unmaximize events to cache the
window preferences. When the window is closed, save them to a Node.js
.config file using configstore. When the app is run, the default window
values are overwritten with the config file. For example, if the window
is resized, maximized, closed, reopened, and then unmaximized, the
window will be sized according to how it was before it was maximized.

Unfortunately... using these Node plugins requires a build process to
bundle them into the Electron main process, so Webpack tooling will be
tweaked in a subsequent commit.

Also, rename app.js to electron.js.